+++ src/flow3r/plugins/core/pipeline/record_audio/pipeline.py
@@ -1,13 +1,9 @@
 from pathlib import Path
-from typing import List, Optional, Dict
+from typing import Optional, Dict
 
 import reactivex as rx
 from reactivex import operators as ops
-from reactivex.disposable import CompositeDisposable
-
-from flow3r.core.api.app.session_context import ISessionContext
-from flow3r.core.pipeline.abc.pipeline import PipelineSubscription, PipelineBase, PreviewSubscription
-from flow3r.core.pipeline.abc.pipeline_type import PipelineType
+from flow3r.core.pipeline.abc.pipeline import ConfigureContext, PreviewContext, PipelineContext, PipelineBase
 from flow3r.core.streaming.abc.stream import IStream
 from flow3r.core.streaming.stream import Stream
 from flow3r.plugins.core.pipeline.record_audio.config import RecordAudioConfig
@@ -18,23 +14,22 @@ class RecordAudioPipeline(PipelineBase[RecordAudioConfig]):
     def __init__(self):
         self._config: Optional[RecordAudioConfig] = None
 
-    def configure(self, session_context: ISessionContext, config: RecordAudioConfig):
-        self._config = config
+    def configure(self, context: ConfigureContext[RecordAudioConfig]) -> None:
+        self._config = context.config
 
-    def preview(self, session_context: ISessionContext, sources: Dict[str, IStream]) -> PreviewSubscription:
-        return PreviewSubscription(CompositeDisposable(), rx.from_list([None]))
+    def preview(self, context: PreviewContext[RecordAudioConfig], sources: Dict[str, IStream]) -> None:
+        context.register_done(rx.from_list([None]))
 
-    def build(self, session_context: ISessionContext, sources: Dict[str, IStream]) -> PipelineSubscription:
+    def build(self, context: PipelineContext[RecordAudioConfig], sources: Dict[str, IStream]) -> None:
         source = sources["Audio"]
 
-        audio_writer_sink = AudioWriterSink(Path(self._config.audio_file))
+        audio_writer_sink = AudioWriterSink(Path(context.config.audio_file))
 
         shared_source = Stream(source.format, source.data.pipe(ops.share()))
         audio_writer_sub = audio_writer_sink.subscribe(shared_source)
 
-        disposable = CompositeDisposable(audio_writer_sub)
-        primary_done = audio_writer_sub.done
-        return PipelineSubscription(disposable, primary_done)
+        context.register_primary_done(audio_writer_sub.done)
+        context.add_disposable(audio_writer_sub.disposable)
 
     def dispose(self):
         pass