+++ src/flow3r/plugins/core/pipeline/record_video_with_audio/pipeline.py
@@ -4,11 +4,9 @@ from typing import Optional, Dict
 
 import reactivex as rx
 from reactivex import operators as ops
-from reactivex.disposable import CompositeDisposable, Disposable
 from reactivex.scheduler import EventLoopScheduler
 
-from flow3r.core.api.app.session_context import ISessionContext
-from flow3r.core.pipeline.abc.pipeline import PipelineSubscription, PipelineBase
+from flow3r.core.pipeline.abc.pipeline import ConfigureContext, PreviewContext, PipelineContext, PipelineBase
 from flow3r.core.streaming.abc.stream import IStream
 from flow3r.core.streaming.stream import Stream
 from flow3r.plugins.core.node.audio_writer_transform import AudioWriterTransform
@@ -24,13 +22,13 @@ class RecordVideoWithAudioPipeline(PipelineBase[RecordVideoWithAudioConfig]):
         self._main_scheduler = EventLoopScheduler()
         self._writer_scheduler = EventLoopScheduler()
 
-    def configure(self, session_context: ISessionContext, config: RecordVideoWithAudioConfig):
-        self._config = config
+    def configure(self, context: ConfigureContext[RecordVideoWithAudioConfig]) -> None:
+        self._config = context.config
 
-    def preview(self, session_context: ISessionContext, sources: Dict[str, IStream]) -> PipelineSubscription:
-        return PipelineSubscription(Disposable(), rx.from_list([None]))
+    def preview(self, context: PreviewContext[RecordVideoWithAudioConfig], sources: Dict[str, IStream]) -> None:
+        context.register_done(rx.from_list([None]))
 
-    def build(self, session_context: ISessionContext, sources: Dict[str, IStream]) -> PipelineSubscription:
+    def build(self, context: PipelineContext[RecordVideoWithAudioConfig], sources: Dict[str, IStream]) -> None:
         video_source = sources["Video"]
         audio_source = sources["Audio"]
 
@@ -40,7 +38,7 @@ class RecordVideoWithAudioPipeline(PipelineBase[RecordVideoWithAudioConfig]):
 
         video_writer_transform = VideoWriterTransform(temp_video_file)
         audio_writer_transform = AudioWriterTransform(temp_audio_file)
-        video_audio_muxer_sink = VideoAudioMuxerSink(Path(self._config.video_file))
+        video_audio_muxer_sink = VideoAudioMuxerSink(Path(context.config.video_file))
 
         shared_video_source = Stream(video_source.format, video_source.data.pipe(ops.observe_on(self._main_scheduler), ops.share()))
 
@@ -56,10 +54,8 @@ class RecordVideoWithAudioPipeline(PipelineBase[RecordVideoWithAudioConfig]):
         )
         muxer_sub = video_audio_muxer_sink.subscribe(video_audio_muxer_stream)
 
-        disposable = CompositeDisposable(muxer_sub)
-        primary_done = muxer_sub.done
-
-        return PipelineSubscription(disposable, primary_done)
+        context.register_primary_done(muxer_sub.done)
+        context.add_disposable(muxer_sub.disposable)
 
     def dispose(self):
         pass