+++ src/flow3r/plugins/core/source/video/pylon/source.py
@@ -19,17 +19,8 @@ class PylonCameraSource(ISource[VideoFormat, VideoFrame]):
             self._video_source.get_fps(),
             "mono8" if self._video_source.get_num_channels() == 1 else "rgb24"
         )
-        last_index = None
-        def check_consecutive(frame: VideoFrame):
-            nonlocal last_index
-            frame_index = frame.frame_index
-            if last_index is None:
-                print(f"First frame index: {frame_index}")
-            elif frame_index - last_index != 1:
-                print(f"Frame index gap: {frame_index - last_index}")
-            last_index = frame_index
 
-        data = reader_observable(self._video_source, read_timeout_seconds=2.0, max_consecutive_errors=30).pipe(ops.do_action(check_consecutive)).pipe(ops.share())
+        data = reader_observable(self._video_source, read_timeout_seconds=2.0, max_consecutive_errors=30).pipe(ops.share())
         self._stream = Stream(fmt, data)
 
     @property